drm: msm: dsi-staging: add support to specify DSI PHY timings in panel dtsi
Some DSI panels require non-standard values of DSI PHY timings for
certain parameters for correct operation. The auto-calculated values
of DSI PHY timings in these cases can show some corruption issues on
the panel. Add support to pass DSI PHY timing parameters from the panel
dtsi in such cases. These value will supersede the PHY timing calculation
logic which is present in the DSI PHY driver.

CRs-Fixed: 2008002

+++ drivers/gpu/drm/msm/dsi-staging/dsi_phy.h
@@ -219,7 +219,7 @@ int dsi_phy_idle_ctrl(struct msm_dsi_phy *phy, bool enable);
  * Return: error code.
  */
 int dsi_phy_set_timing_params(struct msm_dsi_phy *phy,
-			      u8 *timing, u32 size);
+			      u32 *timing, u32 size);
 
 /**
  * dsi_phy_drv_register() - register platform driver for dsi phy
